Add files via upload
this is part of a project that I am working on for class.  We have to add css, but my problem is that in the materials we have when it comes time to add css a blank note pad opens up and all I want to know is how does it link to my project so that any code I write will show up in my final product.